third_party/WebKit/LayoutTests/bluetooth/characteristic/readValue/add-multiple-event-listeners.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  return setBluetoothFakeAdapter('HeartRateAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['heart_rate']}]}))
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryService('heart_rate'))
    .then(service => service.getCharacteristic('body_sensor_location'))
    .then(characteristic => {
      return assert_promise_resolves_after_event(characteristic,
                                                 'readValue',
                                                 'characteristicvaluechanged',
                                                 3 /* attach 3 listeners */);
    }).then(results => {
      let read_value = results[0].buffer;
      let event_values = results.slice(1).map(v => v.buffer);
      for (let event_value of event_values) {
        assert_equals(event_value, read_value);
        assert_array_equals(event_value, read_value);
      }
    });
}, 'Add multiple event listeners then readValue().');
</script>
